+++ src/lib/utils/options/sources/typedoc.ts
@@ -431,6 +431,12 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         type: ParameterType.Boolean,
     });
     options.addDeclaration({
+        name: "markdownLinkExternal",
+        help: (i18n) => i18n.help_markdownLinkExternal(),
+        type: ParameterType.Boolean,
+        defaultValue: true,
+    });
+    options.addDeclaration({
         name: "githubPages",
         help: (i18n) => i18n.help_githubPages(),
         type: ParameterType.Boolean,
@@ -440,7 +446,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
         name: "hostedBaseUrl",
         help: (i18n) => i18n.help_hostedBaseUrl(),
         validate(value, i18n) {
-            if (!/https?:\/\//.test(value)) {
+            if (!/https?:\/\//i.test(value)) {
                 throw new Error(i18n.hostedBaseUrl_must_start_with_http());
             }
         },