+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+-   Added `markdownLinkExternal` option to treat `http[s]://` links in markdown documents and comments as external to be opened in a new tab, #2679.
 -   Added `navigation.excludeReferences` option to prevent re-exports from appearing in the left hand navigation, #2685.
 
 ### Bug Fixes